nemo/requirements/base.txt
django==1.8
python-social-auth
braintree==3.23.0
python==2.7
pycrypto==2.6.1
